Adding tower super class
