+++ settings.py
@@ -34,6 +34,20 @@ TOWER_FIRE_RATE = 250
 TOWER_COST = 250
 TOWER_REFUND = 0.75
 
+# Gun Tower
+GUN_TOWER_DAMAGE = 10
+GUN_TOWER_RADIUS = 200
+GUN_TOWER_FIRE_RATE = 250
+GUN_TOWER_COST = 250
+GUN_TOWER_COLOR = WHITE
+
+# Cannon Tower
+CANNON_TOWER_DAMAGE = 50
+CANNON_TOWER_RADIUS = 400
+CANNON_TOWER_FIRE_RATE = 1000
+CANNON_TOWER_COST = 350
+CANNON_TOWER_COLOR = CYAN
+
 # Mob settings
 MOB_SPEED = 100
 MOB_HEALTH = 100
